requirements.txt

psycopg[binary]==3.2.1
python-dotenv==1.0.1
openai==1.50.2
pypdf==5.0.1
numpy==2.1.0
fastapi==0.115.0
uvicorn==0.30.6
